backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
praw
nltk
vaderSentiment
finnhub-python